requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
alembic==1.16.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.10.1
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1